Cast to (int) in printf, to avoid warning with 64-bit magma_int_t
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@3274 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_zhetrd_gpu.cpp
@@ -61,7 +61,7 @@ int main( int argc, char** argv)
     double tol = opts.tolerance * lapackf77_dlamch("E");
     
     printf("Running version %d; available are (specified through --version num):\n", 
-           opts.version);
+           (int) opts.version);
     printf("1 - uses ZHEMV from CUBLAS (default)\n");
     printf("2 - uses ZHEMV from MAGMA BLAS that requires extra space\n\n");
